New feature: random items allow spawn not themselves, but an item from a brad selection - specify random items in RandomItems.xml - when the game creates a random item, it searches all vialbe items according to random item class and picks one randomly - this allows mappers to place items in maps that are more independent of the xml-set used, however the random items themselves msut be the same in map&mod - can also be used to broaden the enemy gun/item choices selection - for more reference, see http://www.bears-pit.com/board/ubbthreads.php/topics/310793.html#Post310793
FIX: sector curfew value was interpreted wrong

git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@5594 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/XML_SectorNames.cpp
@@ -325,7 +325,7 @@ SectorNameEndElementHandle(void *userData, const XML_Char *name)
 		else if(strcmp(name, "usCurfewValue") == 0)
 		{
 			pData->curElement = SECTORNAME_ELEMENT_SECTOR;
-			pData->usCurfewValue = (UINT16) atoi(pData->szCharData);
+			pData->usCurfewValue = (UINT8) atoi(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
